feat : 예약이 참조하는 테마를 삭제할 경우 예외 발생 기능 추가

+++ src/main/java/roomescape/reservation/service/ThemeService.java
@@ -36,6 +36,7 @@ public class ThemeService {
 
     public List<ThemeResponse> findAll() {
         List<Theme> themes = themeRepository.findAll();
+
         return themes.stream()
                 .map(ThemeResponse::toResponse)
                 .toList();
@@ -49,6 +50,10 @@ public class ThemeService {
     }
 
     public void delete(Long id) {
+        List<Theme> themes = themeRepository.findThemesThatReservationReferById(id);
+        if (!themes.isEmpty()) {
+            throw new IllegalArgumentException("해당 테마로 예약된 내역이 있습니다.");
+        }
         themeRepository.delete(id);
     }
 }
